requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cloudinary==1.42.1
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.0
django-browser-reload==1.15.0
django-cloudinary-storage==0.3.0
django-rest-auth==0.9.5
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==28.1.0
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.1
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.8.2